Unicode character encoding and partial multi-part messaging
Unicode characters are now correctly encoded when messages are sent through the GatorMail Compose page.

Partially integrated the multi-part plaintext/HTML requirement.  That is, there is now an alternative type multipart message, but there is only one part: the plain part.  The HTML part will be added once the FCKeditor is integrated.